add [module] iconArr.js
