+++ script.js
@@ -1,214 +1,7 @@
+import { iconArr } from "./iconArr.js";
+
 const apiKey = "f9321b12d77c24027e5a25c9f625e63b";
-const iconArr = [
-  {
-    name: "clearIcon",
-    svg: `<svg xmlns="http://www.w3.org/2000/svg" width="73.05" height="51.716" viewBox="0 0 73.05 51.716">
-  <g id="noun_Sunglasses_2055147" transform="translate(0 -2.998)">
-    <g id="Group_7" data-name="Group 7" transform="translate(0 2.998)">
-      <g id="Group_6" data-name="Group 6">
-        <g id="Group_1" data-name="Group 1" transform="translate(0 24.387)">
-          <path id="Path_1" data-name="Path 1" d="M17.492,38.339c-3.153,0-5.975-.527-7.725-1.781A15.769,15.769,0,0,1,3.2,25.632c-.04-.24-.064-.426-.1-.551-.28-1-1.72-1.848-2.2-2.067A1.541,1.541,0,0,1,0,21.623V15.536a1.523,1.523,0,0,1,1.059-1.449c19.638-6.291,32.026-1.163,34.3-.073h2.328c2.274-1.087,14.656-6.218,34.3.073a1.524,1.524,0,0,1,1.056,1.449v6.088a1.523,1.523,0,0,1-.9,1.391c-.475.216-1.915,1.065-2.195,2.064-.033.125-.061.31-.1.551a15.768,15.768,0,0,1-6.565,10.927c-3.619,2.6-11.825,2.076-18.062.222-7.034-2.094-8.6-9.944-8.672-10.306h0a.938.938,0,0,0-.03-.125c0,.024-.009.052-.015.079-.07.377-1.631,8.252-8.678,10.349A37.629,37.629,0,0,1,17.492,38.339ZM39.541,25.906c.027.134,1.33,6.4,6.547,7.956,6.864,2.042,13.35,1.7,15.42.222a12.578,12.578,0,0,0,5.336-8.943c.064-.387.119-.688.173-.886a6.562,6.562,0,0,1,2.989-3.54v-4.06c-19.5-5.9-31.119.155-31.238.219a1.529,1.529,0,0,1-.721.183H35a1.551,1.551,0,0,1-.718-.18c-.116-.064-11.837-6.094-31.241-.222v4.06a6.534,6.534,0,0,1,2.986,3.54,7.745,7.745,0,0,1,.173.883,12.572,12.572,0,0,0,5.336,8.943c2.067,1.479,8.553,1.82,15.42-.222,5.25-1.561,6.535-7.9,6.547-7.956a3.086,3.086,0,0,1,2.949-2.773A3.261,3.261,0,0,1,39.541,25.906Z" transform="translate(0 -11.01)" fill="#2a5510"/>
-        </g>
-        <g id="Group_2" data-name="Group 2" transform="translate(48.699)">
-          <path id="Path_2" data-name="Path 2" d="M38.829,33.433a1.52,1.52,0,0,1-1.333-.785L23.233,6.8a1.521,1.521,0,0,0-2.773.356l-1.516,3.978a1.522,1.522,0,0,1-2.843-1.09l1.473-3.853a4.445,4.445,0,0,1,3.8-3.153A4.492,4.492,0,0,1,25.887,5.3L40.159,31.171a1.527,1.527,0,0,1-.6,2.07A1.484,1.484,0,0,1,38.829,33.433Z" transform="translate(-16 -2.998)" fill="#2a5510"/>
-        </g>
-        <g id="Group_3" data-name="Group 3" transform="translate(0.001 0.002)">
-          <path id="Path_3" data-name="Path 3" d="M1.521,33.431a1.521,1.521,0,0,1-1.33-2.258L14.454,5.322a4.568,4.568,0,0,1,8.367.992l1.431,3.726a1.524,1.524,0,0,1-2.846,1.09L19.933,7.276a1.571,1.571,0,0,0-1.324-1.218,1.49,1.49,0,0,0-1.507.758L2.851,32.646A1.512,1.512,0,0,1,1.521,33.431Z" transform="translate(0 -2.999)" fill="#2a5510"/>
-        </g>
-        <g id="Group_4" data-name="Group 4" transform="translate(9.162 30.952)">
-          <path id="Path_4" data-name="Path 4" d="M4.727,21.519a1.522,1.522,0,0,1-1.491-1.227,14.017,14.017,0,0,1-.225-2.615,4.519,4.519,0,0,1,3.869-4.493A1.521,1.521,0,1,1,7.3,16.2a1.487,1.487,0,0,0-1.245,1.467,10.983,10.983,0,0,0,.17,2.039,1.521,1.521,0,0,1-1.2,1.787A1.673,1.673,0,0,1,4.727,21.519Z" transform="translate(-3.01 -13.167)" fill="#2a5510"/>
-        </g>
-        <g id="Group_5" data-name="Group 5" transform="translate(42.635 31.171)">
-          <path id="Path_5" data-name="Path 5" d="M15.744,21.615a1.522,1.522,0,0,1-1.491-1.227,11.957,11.957,0,0,1-.228-2.952,4.523,4.523,0,0,1,3.631-4.167,1.523,1.523,0,1,1,.59,2.989,1.47,1.47,0,0,0-1.184,1.358,8.85,8.85,0,0,0,.177,2.182,1.526,1.526,0,0,1-1.2,1.79A1.616,1.616,0,0,1,15.744,21.615Z" transform="translate(-14.008 -13.239)" fill="#2a5510"/>
-        </g>
-      </g>
-    </g>
-  </g>
-  </svg>
-  `,
-  },
-  {
-    name: "cloudsIcon",
-    svg: `<svg xmlns="http://www.w3.org/2000/svg" width="77.194" height="53.261" viewBox="0 0 77.194 53.261">
-  <g id="noun_Cloud_1188486" transform="translate(-5.8 -20.6)">
-    <g id="Group_8" data-name="Group 8" transform="translate(6.8 21.6)">
-      <path id="Path_13" data-name="Path 13" d="M39.262,72.861a14.468,14.468,0,0,1-12.793-7.746,13.6,13.6,0,1,1-6.527-25.761,18.322,18.322,0,0,1-.174-2.176A15.541,15.541,0,0,1,50.4,33.436a13.065,13.065,0,0,1,6.092-1.392,13.5,13.5,0,0,1,13.49,12.1,12.248,12.248,0,0,1,12.01,11.923A11.931,11.931,0,0,1,64.153,66.42,9.2,9.2,0,0,1,50.228,67.9,14.45,14.45,0,0,1,39.262,72.861ZM26.73,64.071H26.9q.131,0,.261.261A13.378,13.378,0,0,0,39.349,71.99a13.234,13.234,0,0,0,10.618-5.135.452.452,0,0,1,.348-.174c.174,0,.261,0,.348.174A8.351,8.351,0,0,0,63.8,65.55a.538.538,0,0,1,.609-.174,10.569,10.569,0,0,0,5.831,1.654A11.068,11.068,0,0,0,81.3,55.977,11.393,11.393,0,0,0,70.071,44.924h-.435a.411.411,0,0,1-.435-.435,12.726,12.726,0,0,0-18.8-10.1.388.388,0,0,1-.435,0c-.087-.087-.261-.174-.261-.348A14.682,14.682,0,0,0,20.638,37a11.815,11.815,0,0,0,.261,2.524.523.523,0,0,1-.087.348.452.452,0,0,1-.348.174,12.706,12.706,0,1,0-.087,25.413,13.669,13.669,0,0,0,6.092-1.48A.32.32,0,0,0,26.73,64.071Z" transform="translate(-6.8 -21.6)" fill="#f47775" stroke="#f47775" stroke-width="2"/>
-    </g>
-  </g>
-  </svg>`,
-  },
-  {
-    name: "rainIcon",
-    svg: `<svg xmlns="http://www.w3.org/2000/svg" width="73.451" height="73" viewBox="0 0 73.451 73">
-  <g id="noun_Umbrella_2030530" transform="translate(0 -0.435)">
-    <g id="Layer_2" data-name="Layer 2" transform="translate(0 0.435)">
-      <g id="Layer_1" data-name="Layer 1">
-        <path id="Path_6" data-name="Path 6" d="M54.553,83.166a6.28,6.28,0,0,1-6.273-6.273V48.549a1.249,1.249,0,1,1,2.5,0V76.893a3.775,3.775,0,1,0,7.543,0,1.249,1.249,0,0,1,2.5,0,6.28,6.28,0,0,1-6.265,6.273Z" transform="translate(-12.818 -12.955)" fill="#164a68"/>
-        <path id="Path_7" data-name="Path 7" d="M73.451,43.087h-2.2l-.331-.735a6.3,6.3,0,0,0-11.524,0l-.331.735H57.093l-.331-.735a6.3,6.3,0,0,0-11.524,0l-.331.735H42.939l-.323-.735a6.3,6.3,0,0,0-11.532,0l-.323.735H28.793l-.331-.735a6.3,6.3,0,0,0-11.524,0l-.331.735H14.646l-.331-.735a6.3,6.3,0,0,0-11.524,0l-.331.735H0V41.838A34.265,34.265,0,0,1,34.228,7.61h4.995A34.265,34.265,0,0,1,73.451,41.838ZM22.7,36.094a8.814,8.814,0,0,1,7.073,3.57,8.814,8.814,0,0,1,14.147,0,8.814,8.814,0,0,1,14.147,0A8.814,8.814,0,0,1,70.733,38.1a31.775,31.775,0,0,0-31.51-28H34.228A31.782,31.782,0,0,0,2.688,38.335,8.814,8.814,0,0,1,15.63,39.664a8.814,8.814,0,0,1,7.073-3.57Z" transform="translate(0 -2.418)" fill="#164a68"/>
-        <path id="Path_8" data-name="Path 8" d="M49.578,3.87h0a2.248,2.248,0,0,1,2.248,2.248V7.241h-4.5V6.118A2.248,2.248,0,0,1,49.578,3.87Z" transform="translate(-12.566 -1.425)" fill="#164a68"/>
-        <path id="Path_9" data-name="Path 9" d="M20.8,42.047A1.249,1.249,0,0,1,19.549,40.8,36.351,36.351,0,0,1,30.838,13.923,38.283,38.283,0,0,1,42.157,6.3a1.25,1.25,0,0,1,.911,2.328h0A36.373,36.373,0,0,0,32.5,15.8a33.458,33.458,0,0,0-10.415,25A1.249,1.249,0,0,1,20.8,42.047Z" transform="translate(-5.189 -2.047)" fill="#164a68"/>
-        <path id="Path_10" data-name="Path 10" d="M70.476,42.037a1.249,1.249,0,0,1-1.249-1.249c0-23.13-20.838-32.1-21.051-32.186a1.249,1.249,0,1,1,.962-2.3,39.855,39.855,0,0,1,11.3,7.823A36.468,36.468,0,0,1,71.732,40.788,1.256,1.256,0,0,1,70.476,42.037Z" transform="translate(-12.567 -2.036)" fill="#164a68"/>
-        <path id="Path_11" data-name="Path 11" d="M39.875,42.008a1.249,1.249,0,0,1-1.249-1.183c-.044-.948-1.131-23.365,7.7-33.134a1.249,1.249,0,0,1,1.851,1.675c-8.138,9.005-7.088,31.106-7.051,31.327a1.249,1.249,0,0,1-1.183,1.315Z" transform="translate(-10.238 -2.338)" fill="#164a68"/>
-        <path id="Path_12" data-name="Path 12" d="M56.484,41.987H56.41a1.249,1.249,0,0,1-1.175-1.315c0-.22,1.072-22.329-7.051-31.327a1.249,1.249,0,1,1,1.851-1.675c8.814,9.769,7.742,32.186,7.69,33.134a1.249,1.249,0,0,1-1.241,1.183Z" transform="translate(-12.7 -2.317)" fill="#164a68"/>
-        <rect id="Rectangle_4" data-name="Rectangle 4" width="73" height="73" fill="none"/>
-      </g>
-    </g>
-  </g>
-  </svg>
-  `,
-  },
-  {
-    name: "snowIcon",
-    svg: `<svg fill="#fff" height="73" width="73" version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" 
-  viewBox="0 0 512 512" xml:space="preserve">
-  <g transform="translate(0 1)">
-  <g>
-   <g>
-     <path d="M226.025,485.4c5.12,0,8.533-3.413,8.533-8.533v-8.533c0-5.12-3.413-8.533-8.533-8.533c-5.12,0-8.533,3.413-8.533,8.533
-       v8.533C217.492,481.987,220.905,485.4,226.025,485.4z"/>
-     <path d="M260.158,485.4c5.12,0,8.533-3.413,8.533-8.533v-8.533c0-5.12-3.413-8.533-8.533-8.533c-5.12,0-8.533,3.413-8.533,8.533
-       v8.533C251.625,481.987,255.038,485.4,260.158,485.4z"/>
-     <path d="M294.292,485.4c5.12,0,8.533-3.413,8.533-8.533v-8.533c0-5.12-3.413-8.533-8.533-8.533c-5.12,0-8.533,3.413-8.533,8.533
-       v8.533C285.758,481.987,289.172,485.4,294.292,485.4z"/>
-     <path d="M328.425,485.4c5.12,0,8.533-3.413,8.533-8.533v-8.533c0-5.12-3.413-8.533-8.533-8.533s-8.533,3.413-8.533,8.533v8.533
-       C319.892,481.987,323.305,485.4,328.425,485.4z"/>
-     <path d="M362.558,485.4c5.12,0,8.533-3.413,8.533-8.533v-8.533c0-5.12-3.413-8.533-8.533-8.533s-8.533,3.413-8.533,8.533v8.533
-       C354.025,481.987,357.438,485.4,362.558,485.4z"/>
-     <path d="M302.825-1c-70.827,0-128,57.173-128,128v95.573c-7.68-16.213-17.92-31.573-28.16-41.813
-       c-13.653-13.653-40.96-14.507-56.32-1.707c-4.267,4.267-18.773,19.627,0.853,46.933c17.92,25.6,32.427,81.92,32.427,114.347
-       c0,68.385,44.328,99.121,59.733,107.576v37.49c0,14.507,11.093,25.6,25.6,25.6h170.667c14.507,0,25.6-11.093,25.6-25.6v-39.558
-       c25.596-34.759,25.6-38.832,25.6-105.509v-212.48V127C430.825,56.173,373.652-1,302.825-1z M388.158,485.4
-       c0,5.12-3.413,8.533-8.533,8.533H208.958c-5.12,0-8.533-3.413-8.533-8.533v-34.133h187.733V485.4z M413.758,127v213.333
-       c0,63.147,0,64.853-21.333,93.867H193.598c-8.533-4.267-52.907-29.867-52.907-93.867c0-31.573-13.653-93.867-35.84-124.587
-       c-5.973-7.68-11.093-17.92-3.413-23.893c8.533-6.827,25.6-6.827,33.28,0.853c21.333,21.333,40.107,66.56,40.107,87.893
-       c0,5.12,3.413,8.533,8.533,8.533s8.533-3.413,8.533-8.533V127c0-61.44,49.493-110.933,110.933-110.933
-       c60.587,0,110.933,49.493,110.933,110.08V127z"/>
-     <path d="M377.918,191c2.56-4.267,1.707-9.387-2.56-11.947s-9.387-1.707-11.947,2.56L348.62,203.8h-12.75
-       c-3.105-11.949-12.563-21.408-24.512-24.512v-12.75l22.187-14.791c4.267-2.56,5.12-7.68,2.56-11.947s-7.68-5.12-11.947-2.56
-       l-12.8,8.533V127c0-5.12-3.413-8.533-8.533-8.533s-8.533,3.413-8.533,8.533v17.92l-12.8-8.533
-       c-4.267-2.56-9.387-1.707-11.947,2.56s-1.707,9.387,2.56,11.947l22.187,14.791v13.603c-11.949,3.105-21.408,12.563-24.512,24.512
-       H256.46l-14.222-21.333c-2.56-4.267-7.68-5.12-11.947-2.56c-4.267,2.56-5.12,7.68-2.56,11.947l7.964,11.947h-18.204
-       c-5.12,0-8.533,3.413-8.533,8.533c0,5.12,3.413,8.533,8.533,8.533h17.351l-7.964,11.947c-2.56,4.267-1.707,9.387,2.56,11.947
-       c1.707,0.853,3.413,1.707,5.12,1.707c2.56,0,5.12-1.707,6.827-4.267l14.222-21.333h14.172
-       c3.105,11.949,12.563,21.408,24.512,24.512v14.172l-21.333,14.222c-4.267,2.56-5.12,7.68-2.56,11.947
-       c1.707,1.707,4.267,3.413,6.827,3.413c1.707,0,3.413-0.853,5.12-0.853l11.947-7.964v17.351c0,5.12,3.413,8.533,8.533,8.533
-       s8.533-3.413,8.533-8.533v-18.204l11.947,7.964c1.707,0.853,3.413,1.707,5.12,1.707c2.56,0,5.12-1.707,6.827-4.267
-       c2.56-4.267,1.707-9.387-2.56-11.947l-21.333-14.222v-13.319c11.949-3.105,21.408-12.563,24.512-24.512h13.603l14.791,22.187
-       c1.707,1.707,4.267,3.413,6.827,3.413c1.707,0,3.413-0.853,5.12-0.853c4.267-2.56,5.12-7.68,2.56-11.947l-8.533-12.8h17.92
-       c5.12,0,8.533-3.413,8.533-8.533c0-5.12-3.413-8.533-8.533-8.533h-18.773L377.918,191z M302.825,229.4
-       c-9.387,0-17.067-7.68-17.067-17.067c0-9.387,7.68-17.067,17.067-17.067c9.387,0,17.067,7.68,17.067,17.067
-       C319.892,221.72,312.212,229.4,302.825,229.4z"/>
-   </g>
-  </g>
-  </g>
-  </svg>`,
-  },
-  {
-    name: "drizzleIcon",
-    svg: `<svg fill="#004f22" height="73" width="73" version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" 
-    viewBox="0 0 512 512" xml:space="preserve">
-    <g>
-    <g>
-     <path d="M410.071,502.251L371.89,240.836c-2.67-18.313-11.146-35.056-24.119-47.992c17.145-21.023,26.469-47.083,26.469-74.579
-       c0-31.59-12.302-61.288-34.638-83.626C317.264,12.302,287.565,0,255.977,0c-31.59,0-61.288,12.302-83.626,34.639
-       c-43.214,43.214-45.928,111.826-8.144,158.232c-12.959,12.932-21.428,29.665-24.098,47.962l-38.181,261.418
-       c-0.358,2.447,0.367,4.93,1.985,6.802c1.617,1.872,3.969,2.947,6.443,2.947h291.286c2.474,0,4.825-1.075,6.443-2.947
-       C409.705,507.181,410.429,504.699,410.071,502.251z M184.396,46.684c19.12-19.119,44.542-29.649,71.581-29.649
-       s52.46,10.53,71.58,29.649c19.119,19.12,29.649,44.541,29.649,71.58c0,27.039-10.529,52.46-29.649,71.58l-63.062,63.062v-24.091
-       l51.016-51.016c20.921-20.921,29.36-51.744,22.023-80.442c-2.739-10.714-10.055-19.636-20.071-24.479
-       c-10.392-5.024-22.448-5.187-33.074-0.447c-8.959,3.997-18.519,6.024-28.413,6.024c-9.894,0-19.454-2.026-28.413-6.022
-       c-10.629-4.741-22.685-4.579-33.075,0.447c-10.015,4.844-17.331,13.767-20.069,24.479c-7.337,28.697,1.102,59.521,22.023,80.442
-       l51.017,51.017v24.09l-63.062-63.062C144.927,150.375,144.927,86.154,184.396,46.684z M247.459,494.965h-127.25l36.757-251.671
-       c2.107-14.435,8.728-27.651,18.853-37.937l71.64,71.64V494.965z M208.487,165.753c-16.687-16.689-23.419-41.28-17.564-64.177
-       c1.515-5.925,5.415-10.671,10.982-13.363c3.013-1.457,6.279-2.187,9.562-2.187c3.104,0,6.223,0.653,9.156,1.961
-       c11.158,4.976,23.052,7.5,35.353,7.5c12.3,0,24.195-2.523,35.352-7.501c6.031-2.692,12.853-2.61,18.717,0.226
-       c5.567,2.693,9.468,7.439,10.983,13.363c5.854,22.898-0.876,47.489-17.564,64.177l-47.488,47.49L208.487,165.753z
-        M264.494,494.965V276.997l71.666-71.666c10.138,10.29,16.767,23.514,18.874,37.964l36.758,251.67H264.494z"/>
-    </g>
-    </g>
-    <g>
-    <g>
-     <path d="M366.875,430.365l-25.551-61.322c-1.811-4.342-6.798-6.397-11.139-4.586c-4.342,1.809-6.395,6.796-4.586,11.139
-       l25.551,61.322c1.363,3.271,4.529,5.243,7.866,5.243c1.091,0,2.202-0.211,3.273-0.658
-       C366.631,439.695,368.684,434.708,366.875,430.365z"/>
-    </g>
-    </g>
-    <g>
-    <g>
-     <path d="M181.818,364.457c-4.342-1.809-9.328,0.243-11.139,4.586l-25.551,61.322c-1.809,4.343,0.244,9.329,4.586,11.139
-       c1.071,0.446,2.18,0.658,3.273,0.658c3.335,0,6.503-1.973,7.866-5.243l25.551-61.322
-       C188.213,371.253,186.16,366.266,181.818,364.457z"/>
-    </g>
-    </g>
-    </svg>`,
-  },
-  {
-    name: "thunderstormIcon",
-    svg: `<svg fill="#000000" height="73" width="73" version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" 
-    viewBox="0 0 496 496" xml:space="preserve">
-    <g>
-    <g>
-     <g>
-       <path d="M433.896,384H448v-16h-12.248C464.32,328.576,480,280.936,480,232C480,104.072,375.928,0,248,0S16,104.072,16,232
-         c0,48.936,15.68,96.576,44.248,136H48v16h14.104l-20.576,96H24v16h448v-16h-17.528L433.896,384z M32,232
-         c0-119.104,96.896-216,216-216s216,96.896,216,216c0,49.52-17.232,97.624-48.4,136H384V240h16v-36.944l-152-76l-152,76V240h16
-         v128H80.4C49.232,329.624,32,281.52,32,232z M248,256c-26.472,0-48,21.528-48,48v64h-72V240h240v128h-72v-64
-         C296,277.528,274.472,256,248,256z M280,304v64h-64v-64c0-17.648,14.352-32,32-32S280,286.352,280,304z M112,224v-11.056l136-68
-         l136,68V224H112z M57.896,480l3.432-16H88v-16H64.752l6.856-32H88v-16H75.04l3.432-16h339.064l3.432,16H408v16h16.392l6.856,32
-         H408v16h26.672l3.432,16H57.896z"/>
-       <path d="M128,464h240v-64H128V464z M144,416h208v32H144V416z"/>
-       <rect x="216" y="32" width="16" height="16"/>
-       <rect x="256" y="64" width="16" height="16"/>
-       <rect x="320" y="104" width="16" height="16"/>
-       <rect x="336" y="48" width="16" height="16"/>
-       <rect x="200" y="80" width="16" height="16"/>
-       <rect x="144" y="48" width="16" height="16"/>
-       <rect x="80" y="136" width="16" height="16"/>
-       <rect x="128" y="120" width="16" height="16"/>
-       <rect x="392" y="152" width="16" height="16"/>
-       <rect x="432" y="208" width="16" height="16"/>
-     </g>
-    </g>
-    </g>
-    </svg>`,
-  },
-  {
-    name: "otherIcon",
-    svg: `<svg fill="#fffc9c" version="1.1" id="Capa_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" 
-    width="73px" height="73px" viewBox="0 0 482.732 482.732"
-    xml:space="preserve">
- <g>
-   <g>
-     <path d="M374.038,0.009c-4.921-0.173-8.252,2.148-10.034,5.398c-0.538,0.381-1.082,0.772-1.595,1.28
-       c-60.61,59.184-114.688,124.293-170.082,188.256c-13.348-9.105-27.31-16.966-41.787-23.176c-4.535-1.945-8.844-1.062-12.137,1.198
-       c-1.274,0.117-2.574,0.365-3.905,0.914c-25.89,10.796-46.819,28.543-62.914,51.196c-3.052,1.315-5.261,3.95-6.437,6.972
-       c-2.872,2.534-4.42,6.424-2.912,11.299c24.075,77.794,90.799,185.29,180.715,192.866c4.271,0.36,7.3-1.254,9.188-3.752
-       c1.676-0.198,3.402-0.717,5.175-2.037c25.258-18.778,44.544-42.081,55.35-71.853c0.625-1.717,0.823-3.306,0.727-4.764
-       c0.229-1.665,0.065-3.499-0.727-5.535c-5.093-13.096-11.263-26.151-18.306-38.953c58.387-60.986,113.264-125.278,173.769-184.244
-       c2.163-2.107,3.229-4.453,3.498-6.769c2.793-1.899,4.9-4.905,5.286-9.085C482.803,45.478,435.669,2.137,374.038,0.009z
-        M90.677,252.765c19.4,6.546,37.333,16.321,53.458,28.473c-3.864,19.251-4.778,41.197,7.482,57.625
-       c12.667,16.966,33.55,18.819,52.857,15.224c8.513,17.082,14.79,35.333,18.278,54.259
-       C160.811,389.51,112.604,311.939,90.677,252.765z M162.611,296.99c11.75,11.324,22.208,24.009,31.011,37.811
-       c-9.476,1-19.39-0.203-25.375-8.216C162.136,318.388,161.486,307.419,162.611,296.99z M246.832,408.568
-       c-14.097-81.085-70.376-152.705-148.896-178.729c12.682-15.056,28.417-26.507,46.834-34.438
-       c63.975,29.828,118.832,93.09,144.375,158.042C280.857,375.617,265.582,394.299,246.832,408.568z M453.112,106.652
-       c-0.646,0.432-1.285,0.919-1.899,1.519c-59.067,57.563-112.792,120.19-169.635,179.877
-       c-19.073-29.473-42.919-56.731-70.089-78.785C250.765,163.85,289.484,117.94,330.362,74l19.362,19.367
-       c10.912,10.908,27.827-6.007,16.914-16.914l-19.834-19.84c10.455-10.923,21.022-21.754,31.849-32.347
-       C424.828,27.771,456.286,59.122,453.112,106.652z" fill="#fffc9c" stroke="#fffc9c"/>
-     <path d="M19.76,340.686c12.042-4.058,23.899-8.633,35.942-12.685c14.563-4.911,8.554-28.071-6.068-23.146
-       c-12.042,4.052-23.899,8.632-35.942,12.685C-0.879,322.451,5.138,345.611,19.76,340.686z"fill="#fffc9c" stroke="#fffc9c"/>
-     <path d="M89.666,421.085l25.372-25.369c10.908-10.908-6.007-27.822-16.917-16.915l-25.37,25.369
-       C61.849,415.083,78.759,431.993,89.666,421.085z"fill="#fffc9c" stroke="#fffc9c"/>
-     <path d="M170.342,470.773c1.077-10.582,4.202-20.921,8.072-30.798c2.374-6.073-2.793-13.101-8.539-14.604
-       c-6.85-1.788-12.225,2.462-14.606,8.536c-4.674,11.938-7.556,24.12-8.846,36.865c-0.66,6.495,5.938,11.959,11.961,11.959
-       C165.392,482.732,169.688,477.248,170.342,470.773z" fill="#fffc9c" stroke="#fffc9c"/>
-     <path d="M290.394,226.949c18.098,18.098,46.148-9.963,28.061-28.056C300.356,180.8,272.301,208.862,290.394,226.949z"fill="#fffc9c" stroke="#fffc9c"/>
-   </g>
- </g>
- </svg>`,
-  },
-];
+
 let currentLocation = null;
 let currentDt = null;
 let currentWeather = null;
@@ -358,14 +151,23 @@ const getSunTime = (weatherData) => {
   const { sunset } = weatherData.sys;
   const { timezone } = weatherData;
 
+  // Get the local timezone offset in seconds
+  const localTimezoneOffset = new Date().getTimezoneOffset() * 60;
+
   //convert UTC timeStamp into human-readable data
-  const sunriseData = new Date((sunrise + timezone) * 1000);
-  const sunsetData = new Date((sunset + timezone) * 1000);
+  const sunriseData = new Date(
+    (sunrise + timezone + localTimezoneOffset) * 1000
+  );
+  const sunsetData = new Date((sunset + timezone + localTimezoneOffset) * 1000);
 
   //extract the exact time of the day by converting to local time
   //to prevent winter/summer time problem
   const formatTime = (date) => {
-    return date.toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" });
+    return date.toLocaleTimeString([], {
+      hour: "2-digit",
+      minute: "2-digit",
+      hour12: false,
+    });
   };
 
   const sunriseTime = formatTime(sunriseData);
@@ -491,11 +293,13 @@ const searchWeather = () => {
   getDailyWeather(userSearch);
   getWeeklyWeather(userSearch);
 };
+//add event listeners
 document.querySelector(".searchBtn").addEventListener("click", searchWeather);
 document.querySelector(".search-bar").addEventListener("keyup", (event) => {
   if (event.key === "Enter") {
     searchWeather();
-    console.log(userSearch);
+    const targetElement = document.querySelector(".tip");
+    targetElement.scrollIntoView();
   }
 });
